cpp: Switch build system to CMake
This sets up a better structure and build system for the project, and
makes it easier to use it as a target in other projects.

Build time options have also been included, where static and shared
library modes are supported. For now, static library mode does not work
when tests are enabled, as Fedora doesn't seem to ship a static version
of the googletest library, but this is a problem of the build
environment (and users of the library will not build the tests anyway).
Eventually we may switch to submodule for googletest.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>